+++ lib/generator.rb
@@ -29,7 +29,6 @@ class Generator
     split_key_hash[:C] = keys[2] + keys[3]
     split_key_hash[:D] = keys[3] + keys[4]
     split_key_hash.values
-    #returns strings of values
   end
 
   def generate_offset(date)
@@ -38,7 +37,6 @@ class Generator
       number.to_i
     end
     offsets
-    #returns integers
   end
 
   def shift(keys, offsets)
@@ -48,53 +46,7 @@ class Generator
     shift_hash[:C] = keys[2].to_i + offsets[2]
     shift_hash[:D] = keys[3].to_i + offsets[3]
     shift_hash
-    #returns hash with integers
   end
-
-  def encrypted_string(string, option_key = generate_keys, option_date = date)
-    given_key = split_keys(option_key)
-    given_date = generate_offset(option_date)
-    
-    shift_numbers = shift(given_key, given_date).values 
-    shifted_character_collector = ''
-    string.downcase.split('').each do |character|
-      if !letters.include?(character)
-        shifted_character_collector += character
-      else
-        value_shift = letters.find_index(character) + shift_numbers.first
-        while value_shift >= letters.length
-          value_shift -= letters.length
-        end
-        the_shift = letters[value_shift]
-        shifted_character_collector += the_shift 
-        shift_numbers.rotate!
-      end
-    end
-    shifted_character_collector
-  end
-
-  def decrypted_string(string, option_key, option_date = date)
-    given_key = split_keys(option_key)
-    given_date = generate_offset(option_date)
-
-    shift_numbers = shift(given_key, given_date).values 
-    shifted_character_collector = ''
-    string.downcase.split('').each do |character|
-      if !letters.include?(character)
-        shifted_character_collector += character
-      else
-        value_shift = letters.find_index(character) - shift_numbers.first
-        while value_shift < 0
-          value_shift += letters.length
-        end
-        the_shift = letters[value_shift]
-      
-        shifted_character_collector += the_shift 
-        shift_numbers.rotate!
-      end
-    end
-    shifted_character_collector
-  end 
   
   def date
     Date.today.strftime('%d%m%y')
